mark SDD0023 obsolete
We figured out that this concept is not needed in this form.
By using the modern Crossplane concepts we can achieve similar things.
Plus in the mean-time there are similar controllers available which
could be reused if we still would need similar concepts as discussed in
this SDD, e.g. https://kyverno.io/